plugins/elab-tutor/elab-local/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
ollama>=0.4.0
pyyaml>=6.0
